change tools file and put Generator in own file.
